mychatbackend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.4
langsmith==0.1.117
MarkupSafe==2.1.5
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.44.1
orjson==3.10.7
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.4
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.34
tenacity==8.5.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.4
yarl==1.11.1
